+++ src/components/PDF/PDF.tsx
@@ -16,6 +16,16 @@ import { CMSData } from '../../cms-integration/getCMSIntegration';
 import { getFullName } from '../../helpers/utils';
 import accents from '../../tokens/accents';
 import neutrals from '../../tokens/neutrals';
+import { htmlRenderers } from './htmlRenderers';
+import { BuildingColumns } from './Icons/BuildingColumns';
+import { Calendar } from './Icons/Calendar';
+import { CircleBriefcase } from './Icons/CircleBriefcase';
+import { CircleCheck } from './Icons/CircleCheck';
+import { CircleGraduationCap } from './Icons/CircleGraduationCap';
+import { CircleIdCard } from './Icons/CircleIdCard';
+import { CirclePaintbrush } from './Icons/CirclePaintbrush';
+import { CircleUser } from './Icons/CircleUser';
+import { Star } from './Icons/Star';
 
 const accentColor = accents[resumeConfig.accentColor].light;
 const neutralColor = neutrals[resumeConfig.neutralColor].light;
@@ -23,44 +33,56 @@ const neutralColor = neutrals[resumeConfig.neutralColor].light;
 const domain = process.env.VERCEL_URL
   ? `https://${process.env.VERCEL_URL}`
   : 'http://localhost:3000';
-const fontPath = `${domain}/fonts/SourceSansPro`;
+const fontPath = `${domain}/fonts`;
 const iconPath = `${domain}/pdf/fa-icons`;
 
 Font.register({
-  family: 'Source Sans Pro',
+  family: 'Albert Sans',
   fonts: [
     {
       fontStyle: 'normal',
       fontWeight: 400,
-      src: `${fontPath}/SourceSansPro-Regular.ttf`,
+      src: `${fontPath}/AlbertSans-Regular-Converted.ttf`,
     },
     {
       fontStyle: 'italic',
       fontWeight: 400,
-      src: `${fontPath}/SourceSansPro-Italic.ttf`,
+      src: `${fontPath}/AlbertSans-Italic-Converted.ttf`,
     },
     {
       fontStyle: 'normal',
       fontWeight: 700,
-      src: `${fontPath}/SourceSansPro-Bold.ttf`,
+      src: `${fontPath}/AlbertSans-Bold-Converted.ttf`,
     },
     {
       fontStyle: 'italic',
       fontWeight: 700,
-      src: `${fontPath}/SourceSansPro-BoldItalic.ttf`,
+      src: `${fontPath}/AlbertSans-BoldItalic-Converted.ttf`,
     },
   ],
 });
 
-const sidebarWidth = 2.75;
+const hyphenationCallback = (word) => {
+  // don't hyphenate
+  return [word];
+};
+
+Font.registerHyphenationCallback(hyphenationCallback);
+
+Font.registerEmojiSource({
+  format: 'png',
+  url: 'https://cdnjs.cloudflare.com/ajax/libs/twemoji/14.0.2/72x72/',
+});
+
 const fontSizes = {
   xl: 20,
   l: 18,
-  m: 16,
-  s: 14,
+  m: 14,
+  s: 13,
   xs: 12,
   xxs: 10,
 };
+
 const spacers = {
   1: '6px',
   2: '8px',
@@ -77,7 +99,7 @@ const styles = StyleSheet.create({
     display: 'flex',
     flexDirection: 'row',
     flexWrap: 'nowrap',
-    fontFamily: 'Source Sans Pro',
+    fontFamily: 'Albert Sans',
     fontSize: fontSizes.xxs,
     justifyContent: 'flex-start',
     lineHeight: 1.3,
@@ -87,7 +109,7 @@ const styles = StyleSheet.create({
     backgroundColor: neutralColor[3],
     display: 'flex',
     color: neutralColor[12],
-    flexBasis: `${sidebarWidth}in`,
+    flexBasis: '30%',
     flexDirection: 'column',
     flexGrow: 0,
     flexShrink: 1,
@@ -100,11 +122,11 @@ const styles = StyleSheet.create({
     textAlign: 'center',
   },
   headerTitle: { fontSize: fontSizes.xl, fontWeight: 700 },
-  headerSubtitle: { fontSize: fontSizes.l, fontWeight: 700 },
+  headerSubtitle: { fontSize: fontSizes.m, fontWeight: 700 },
   main: {
     alignSelf: 'stretch',
     display: 'flex',
-    flexBasis: 'auto',
+    flexBasis: '70%',
     flexDirection: 'column',
     flexGrow: 1,
     flexShrink: 0,
@@ -117,6 +139,7 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     fontSize: fontSizes.m,
     fontWeight: 700,
+    gap: spacers[1],
   },
   sectionHeadingNonHTML: {
     alignItems: 'center',
@@ -124,6 +147,7 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
     fontSize: fontSizes.m,
     fontWeight: 700,
+    gap: spacers[1],
     marginBottom: spacers[1],
   },
   sectionHeadingIcon: {
@@ -135,19 +159,18 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     display: 'flex',
     flexDirection: 'row',
-    marginRight: spacers[1],
   },
-  sectionHeadingStar: {
-    height: fontSizes.xxs,
-    width: 'auto',
+  sectionParagraph: {
+    fontWeight: 400,
+    margin: 0,
   },
-  sectionParagraph: { fontWeight: 400, margin: 0 },
   itemHeading: {
     alignItems: 'center',
     display: 'flex',
     flexDirection: 'row',
     fontSize: fontSizes.s,
     fontWeight: 700,
+    gap: spacers[1],
     marginBottom: spacers[1],
     marginTop: spacers[3],
   },
@@ -155,15 +178,12 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     display: 'flex',
     flexDirection: 'row',
+    gap: spacers[1],
     marginBottom: spacers[1],
   },
   itemSubheading: {
     fontSize: fontSizes.xxs,
     fontStyle: 'italic',
-  },
-  itemSubheadingIcon: {
-    height: fontSizes.xxs,
-    marginRight: spacers[1],
   },
   professionalTitle: {
     backgroundColor: neutralColor[12],
@@ -184,13 +204,19 @@ const styles = StyleSheet.create({
     color: accentColor[11],
     textDecoration: 'underline',
   },
+  list: {
+    marginTop: spacers[2],
+  },
 });
 
 const htmlProps: Omit<HtmlProps, 'children'> = {
+  renderers: htmlRenderers,
   style: { fontSize: fontSizes.xxs },
   stylesheet: {
     a: styles.a,
     p: styles.sectionParagraph,
+    ul: styles.list,
+    ol: styles.list,
   },
 };
 
@@ -221,20 +247,14 @@ const PDF: React.FC<CMSData> = (props) => {
           <View style={styles.sidebarContent}>
             <View style={styles.section}>
               <View style={styles.sectionHeadingNonHTML}>
-                <Image
-                  src={`${iconPath}/circle-user.png`}
-                  style={styles.sectionHeadingIcon}
-                />
+                <CircleUser size={fontSizes.m} />
                 <Text>About Me</Text>
               </View>
               <Html {...htmlProps}>{personalInformation.html}</Html>
             </View>
             <View style={styles.section}>
               <View style={styles.sectionHeadingNonHTML}>
-                <Image
-                  src={`${iconPath}/circle-id-card.png`}
-                  style={styles.sectionHeadingIcon}
-                />
+                <CircleIdCard size={fontSizes.m} />
                 <Text>Contact Information</Text>
               </View>
               <View style={styles.flexRow}>
@@ -252,10 +272,7 @@ const PDF: React.FC<CMSData> = (props) => {
             </View>
             <View style={styles.section}>
               <View style={styles.sectionHeading}>
-                <Image
-                  src={`${iconPath}/circle-checkmark.png`}
-                  style={styles.sectionHeadingIcon}
-                />
+                <CircleCheck size={fontSizes.m} />
                 <Text>Skills &amp; Expertise</Text>
               </View>
               {skills.map((skill, skillIndex) => (
@@ -264,11 +281,7 @@ const PDF: React.FC<CMSData> = (props) => {
                     <View style={styles.sectionHeadingStars}>
                       {Array.from(Array(skills.length - skillIndex)).map(
                         (star, starIndex) => (
-                          <Image
-                            key={starIndex}
-                            src={`${iconPath}/star-yellow.png`}
-                            style={styles.sectionHeadingStar}
-                          />
+                          <Star key={starIndex} size={fontSizes.xxs} />
                         ),
                       )}
                     </View>
@@ -283,10 +296,7 @@ const PDF: React.FC<CMSData> = (props) => {
         <View style={styles.main}>
           <View style={styles.section}>
             <View style={styles.sectionHeading}>
-              <Image
-                src={`${iconPath}/circle-briefcase.png`}
-                style={styles.sectionHeadingIcon}
-              />
+              <CircleBriefcase size={fontSizes.m} />
               <Text>Professional Experience</Text>
             </View>
             {professional.map((professionalExperience) => (
@@ -300,10 +310,7 @@ const PDF: React.FC<CMSData> = (props) => {
                   </Text>
                 </View>
                 <View style={styles.itemSubheadingRow}>
-                  <Image
-                    src={`${iconPath}/calendar.png`}
-                    style={styles.itemSubheadingIcon}
-                  />
+                  <Calendar size={fontSizes.xxs} />
                   <Text style={styles.itemSubheading}>
                     {professionalExperience.attributes.startDate}—
                     {professionalExperience.attributes.endDate
@@ -317,10 +324,7 @@ const PDF: React.FC<CMSData> = (props) => {
           </View>
           <View style={styles.section}>
             <View style={styles.sectionHeading}>
-              <Image
-                src={`${iconPath}/circle-graduation-cap.png`}
-                style={styles.sectionHeadingIcon}
-              />
+              <CircleGraduationCap size={fontSizes.m} />
               <Text>Achievements</Text>
             </View>
             {achievements.map((achievement) => (
@@ -331,10 +335,7 @@ const PDF: React.FC<CMSData> = (props) => {
                   </Text>
                 </View>
                 <View style={styles.itemSubheadingRow}>
-                  <Image
-                    src={`${iconPath}/university.png`}
-                    style={styles.itemSubheadingIcon}
-                  />
+                  <BuildingColumns size={fontSizes.xxs} />
                   <Text style={styles.itemSubheading}>
                     {achievement.attributes.institution}
                   </Text>
@@ -345,10 +346,7 @@ const PDF: React.FC<CMSData> = (props) => {
           </View>
           <View style={styles.section}>
             <View style={styles.sectionHeading}>
-              <Image
-                src={`${iconPath}/circle-pen-paintbrush.png`}
-                style={styles.sectionHeadingIcon}
-              />
+              <CirclePaintbrush size={fontSizes.m} />
               <Text>Hobbies &amp; Interests</Text>
             </View>
             <Html