+++ CHANGELOG.md
@@ -5,12 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2023-02-22
+
+### Changed
+
+- Dependency upgrades
+- Default to Albert Sans as the PDF font to match the web look and feel
+  - Adjust font sizes accordingly
+- Don't hyphenate paragraphs in PDF
+- Switch to rendering SVG components for PDF icons instead of raster images
+
+### Fixed
+
+- Improved rendering of ordered and unordered lists in PDF
+  - Thanks to [justin-hackin](https://github.com/justin-hackin) for the find and custom renderer function
+- Minor flexbox improvements
+  - Switch to percentage layout of sidebar and main content on PDF instead of basing it on inches
+
 ## [3.0.3] - 2023-02-21
 
 ### Changed
 
 - Dependency upgrades
--
 
 ## [3.0.2] - 2022-12-31